requirements.txt
apturl==0.5.2
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
certifi==2019.11.28
chardet==3.0.4
checkbox-support==0.22
click==6.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
crayons==0.3.0
cryptography==1.2.3
defer==1.0.6
Deprecated==1.2.7
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.9
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
prompt-toolkit==0.57
ptyprocess==0.5
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
Pygments==2.1
pygobject==3.20.0
PyJWT==1.7.1
PyMySQL==0.7.2
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.8
python-debian==0.1.27
python-dotenv==0.12.0
python-systemd==231
pyxdg==0.25
reportlab==3.3.0
requests==2.23.0
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
six==1.10.0
sqlparse==0.1.18
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.25.8
usb-creator==0.3.0
wcwidth==0.1.6
wrapt==1.12.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
